Add files via upload
The objective of this practical is to develop a Python program that 
calculates the area and perimeter of a rectangle based on user input for length and 
width. This program defines the area of rectangle and perimeter of rectangle 
respectively. This exercise will help students understand basic programming concepts 
like input/output operations, data types, and arithmetic operations in Python